refactor: rename mihomo-websocket to managedMihomoWs
