+++ src/hooks/use-log-data.ts
@@ -4,7 +4,7 @@ import type { Log, LogLevel } from "tauri-plugin-mihomo-api";
 import {
   ManagedMihomoWebSocket,
   subscribeManagedMihomoWebSocketText,
-} from "@/services/mihomo-websocket";
+} from "@/services/managedMihomoWs";
 import { mutate, swrSubscriptionKey, useSWRSubscription } from "@/services/swr";
 import { useClashLogStore, useRefreshLogsDateStore } from "@/stores";
